refactor(solver): split function-apparent helpers out of core_dispatch to restore the file-size ceiling
#16483 grew crates/tsz-solver/src/relations/subtype/core_dispatch.rs to 2075
lines — over the 2000-line §19 ceiling — by adding the
`or_global_function_interface_surface` second-opinion helper and the
`function_apparent_object_shape` builder inline. That pushed the solver
oversized-file count past its ratchet baseline (the nightly
`solver_file_size_ceiling` guard).

Per the repo rule "split instead of adding local allowlists or ceilings,"
move both methods verbatim into a new subtype/function_apparent.rs module (a
pure relocation via `use super::*`; behavior unchanged). This brings
core_dispatch back to 1988 lines and restores the solver file-size ceiling.
